backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "alembic-postgresql-enum>=1.7.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "greenlet>=3.1.1",
    "psycopg>=3.2.6",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.399",
    "ruff>=0.11.5",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 80
lint.ignore = ["E501"]

